Worked on notification
